Add free disk space check to health-checker.py. Move load-average check to health-checker.py.
